기존 BVO파일 변수 수정_서

+++ A823/src/main/java/co/sol/service/BService.java
@@ -6,7 +6,7 @@ import co.sol.main.Basic;
 
 public interface BService {
 	
-	Long submit(BVO bo);
+	int submit(BVO bo);
 	
 	BVO get(int bnum);
